Add a simple template egs++ application egs_app
Provide the simplest possible EGSnrc application, which relies on the
implementation of EGS_AdvancedApplication, to run a simulation described
in an egs++ input file (the provided slab.egsinp, for example).

The code commented below APP_MAIN can serve as a template for deriving an
application class from EGS_AdvancedApplication, and re-implementing the
initScoring method to select ausgab call triggers, and the ausgab method
to handle those calls.