Add pulse height distribution application egs_phd
Provide a sample EGSnrc application which scores a pulse height
distribution in a set of regions specified by a label. Included is a
sample input to model a germanium detector inside a lead shield,
irradiated by a collimated point source. Also included are egs_view
settings and a viewer screen shot for reference.