Reduce lock file race condition with random delay
Add a random component to the delay before a thread retries to lock the file
upon failure, to reduce the risk of future collisions with other threads.
Previously, threads waited for a fixed 12 seconds before trying to lock the
file again. Tested on PBS scheduler clusters and with NFS and HDD storage.