Fix ambiguities and bugs in radionuclide source
Account for the uncertainties on the decay intensities when the total
decay intensity does not equal 100. Use the total intensity for internal
conversion, even if the individual shell intensities don't add up to
equal the total.

Improve the handling of internal transitions from unknown levels,
including fixing a bug where the total transition intensity was assigned
to the gamma intensity.

Add the option to model extra transitions when there is a mismatch
between the intensities in and out of a level in the daughter.

Fix a bug where internal pair intensity was read as P-shell internal
conversion intensity. Remove a duplicated decay in the Th-234 ensdf
file.