Fix radionuclide auger emissions for EADL
Fix a bug where the auger electrons from atomic relaxations during
egs_radionuclide_source source generation were not produced. This did
not affect atomic relaxations in EGSnrc in general, only the
radionuclide source when it was set to EADL relaxation mode. The auger
electrons were excluded because the relaxation code expected ECUT to be
provided in kinetic energy instead of total energy.