Fix DOSXYZnrc parallel runs
Fix a bug in DOSXYZnrc parallel runs when they were initiated using the
`Run job in parallel' option in the GUI. The _w tag was not added to the
pardose files, resulting in a crash when the code tried to combine them.
This bug did not affect parallel runs using the -j command line
argument.