Fix #528: Windows PowerShell bug with exe
Fix a Windows PowerShell bug where the .exe extension on applications
was included in the app_name variable. This resulted in a runtime error
where the input file failed to be found. Now we check to see if there is
a .exe at the end of the app_name, and remove it.