Fix test for synchronized modules name
Check if characters 1-4 of the component module name (CMNAME) for any
module in the accelerator are 'SYNC', instead of checking the full
CMNAME string, to enable the scoring of the monitor unit index in IAEA
formal phase space files (I_MU_PHSP). Previously, the test failed when
the only synchronized module was SYNCHDMLC, because the CMNAME string is
only 8 characters long!