Clarify the simple alias table construction
Clarify the algorithm to build a simple alias table. This version is
easier to understand and is better guarded. With the float guard epsilon
on the big bin height comparison the loop is guaranteed to finish, but
we leave the infinite loop guard anyways.

For reference, the algorithm now follows closely the pseudo-code given
in Luc Devroye, Non-Uniform Random Variate Generation (1986).