Fix #554: compile error in BEAMnrc with EEMF
Remove the undeclared variable ofr_cas in the enhanced electromagnetic
field (EEMF) macros, which caused an error when compiling with BEAMnrc.
Remove a couple unused constants and fix a typo in the file name
(beamnrc instead of beamrc). Also update the EEMF copyright details
and add publication reference in file headers.