Add egs_glib to load a geometry from another file
The egs_glib geometry can load a geometry which is defined in an
external file, allowing geometry definitions to be modularized across
multiple files, reused, shared, etc.

There are options to compile with gzip capability (for egs_glib and
egs_autoenvelope), and to compile with a Sobol random number generator
(for the egs_autoenvelope region discovery sampling). For either of
these features, install the egspp-geometry-lib-extras from:

https://github.com/clrp-code/egspp-geometry-lib-extras/