Fix file includes in egs++ input files
Fix file includes in egs++, so that the file content is inserted into
the total input file content before everything else is processed. This
fixes a bug where composite geometries that referenced a geometry in the
external file would crash, because the external geometry hadn't been
loaded and assigned a name yet.