Fix egs_circle_perpendicular particle weights
Fix the particle weights when egs_circle_perpendicular is used as
a target. Previously the weighting was based on the viewing angle
of the circle along the z axis, because the circle was always in the
xy plane in egs_circle. Now, since the circle face changes direction
the weighting is constant with viewing angle.