Label fluorescent photons from eii as secondaries
Label fluorescent photons from electron impact ionization (eii) as
secondaries, and store the bit region of origin in latch bits 24-28 of
fluorescent photons.

With no splitting, this is done using a call to ausgab with iarg=$EIIA
(flag denoting that eii just occurred), looping through npold..np and
labeling all resultant photons.

With UBS and DBS, we label fluorescent photons prior to calling
splitting routines. While this is slightly overkill (it "relabels"
fluorescent photons from Compton and photoelectric relaxations), it is
necessary for eii because of the redefinition of npold in these cases.