Added analogous binary store/setState functions
Added binary analogues to setState and storeState (setBinState and storeBinState) to be implemented
in applications using a data stream seperate from data_out.  For example:

```c
int egs_mird::outputData() {
    // Invoke base class first
    int err = EGS_AdvancedApplication::outputData();
    if (err) {
        return err;
    }
	
    // Create new binary egsdat file (begsdat) for faster data storage
    string ofile = constructIOFileName(".begsdat",true);
    ostream *data_bout = new ofstream(ofile.c_str(), ios::binary | ios::out);
	
    // Store score
    if (!score->storeState(*data_bout)) {
        return 101;
    }

    delete data_bout;
    return 0;
}
```

WARNING:  All data is output as it is stored in memory, and can occupy several times as much space
as analogous text files.