Account for multiple CV entries in egs_kerma
Account for the possibility of multiple entries into the scoring volume
when ray-tracing along the photon's direction when using forced
detection (exponential track-lenght estimation). This is relevant for
hollow geometries such as spherical or cylindrical shells. The scoreInCV
method calls itself recursively if the photon re-enters the FD geometry
or is still inside it.