Add splitting var in egs_chamber to replace latch
Use a dedicated splitting variable nbr_split, instead of latch, to hold
the splitting number in the egs_chamber application. A new stack
variable nbr_splitting is added to pass its value through the EGSnrc
back end. This avoids overwriting the latch variable in cases where it
is needed otherwise, for example in the phsp scoring object. Ideally,
the meaning of latch should remain consistent with BEAMnrc.