Add an egs-parallel sub-script for multicore cpus
This egs-parallel-cpu subscript provides the option "--batch cpu" to
egs-parallel, to launch a simulation on multiple cores on the local cpu,
without requiring a job scheduler. Intentionally, this script is simple:
it just launches the jobs sequentially, without waiting around for the
.egsjob or .lock files, as in the pbs scripts. However, the logging is
consistent with the other egs-parallel scripts

The number of threads is always constrained to the number of threads
available on the machine, because it is inefficient to go beyond that,
and launching a large number of threads on a cpu by mistake may well
stall the computer.