Add -l (--list) option to egs-parallel-clean
For convenience, add a -l (--list) option to the cleaning script to list
all the .egslog file base names in the current directory. This option is
checked first and overrides every other argument: the list is printed to
the terminal and the script terminates. Also, reformat the usage message
and use the extension .egsparallel-eo (with a hyphen) to avoid collision
with the pbs .eo extension. Use executable basename in quit function.