Fix total ncase for the job control file object
Reset ncase to the user requested value by multiplying it by the number
of parallel jobs.

Previously, the job control file object (JCFO) effectively reduced the
total number of histories by the number of parallel jobs. This bug was
introduced in the initial implementation of the uniform RCO where ncase
was reduced in the base class of the RCO. This works for the simple and
uniform types, but not for the JCFO, where the number of histories are
managed by the job control file object (via a lock file).

Also, ensure that by default the last job is a watcher job in the
uniform RCO even if there is no input block defining an RCO in the input
file.