Check uniform RCO job termination in Motran apps
For the uniform run control object (RCO), check if the number of output
files (i.e., *egsdat, *pardose, etc.) is equal to the number of combined
parallel jobs. If not, then the last job sleeps for a set time
($URC-SLEEP, by default 1 second) and repeats the combining process a
set number of times ($URC-INTERVALS, by default only once).

Previously, the uniform RCO for Mortran apps was in fact identical to
the egs++ simple URC. In that case, the last job combines only those
jobs that finished before it.

Note that the Mortran uniform RCO implementation does not delete
existing output files such as *egsdat and *pardose at the beginning of a
parallel run (as opposed to the egs++ implementation). This could
produce ill effects when re-running a previous calculation with a
modified input file: if not all output files are overwritten (e.g.,
fewer parallel jobs are requested for the new run), then a mix of old
and new output files will be combined.