Preserve information after DOSXYZnrc parallel run
Add calls to summarizing subroutines to preserve information in the
combined egslog and egslst files. This way, if the intermediate output
files from individual jobs are removed, this information remains
available. No changes were strictly required in DOSXYZnrc for the
uniform run control object implementation.

Previously, the only information in the final output file pertained to
the total dose in each requested region of the phantom, because in
DOSXYZnrc combining results is implemented for *.pardose files only.

Note that the timing information is missing from the final output file.