Fix timing and atomic relaxations in CAVRZnrc
Account for the overall cpu time to properly compute the number of
histories per hour. Previously, the total time was not included, but the
total number of histories was.

Turn off atomic relaxations in the cavrznrc_template.egsinpsample input
file, otherwise if fails because bound Compton scattering is turned off.
Funnily, it doesn't fail if the photon xsection library SI is used. Why?