EGS_Mesh: remove msh2 support, only support msh4
The original intent for the msh file parser was to support both the
legacy (but stable) v2.2 and today's default (v4.1). Thinking again,
it's better if we only need to maintain a single parser. v4.1 is the
default for Gmsh today. There are no planned changes to the sections we
care about ($Entities, $PhysicalNames, $Elements and $Nodes). Any future
4.x revisions will hopefully not break the parser and we could "support"
them by accepting the new versions in parse_msh_version.