Fix dose values reported in egs++ 3ddose output
Fix the value of dose in 3ddose files generated from egs++ apps using
the egs_dose_scoring ausgab object. The dose was not being scaled by the
mass of materials, because the getMass function in the ND geometry class
was not defined properly for XYZ geometries. This critical bug has
existed since the inception of 3ddose file output in egs_dose_scoring
(in commit 7e5bf079 on 25 July 2017). In the limited case I tested, the
density in every voxel was 1 instead of the proper material density.