Fix command to get file size in configure script
Use 'wc -c' to obtain the number of bytes in the temporary file when
determining the Fortran record length during EGSnrc configuration. The
problem with 'stat -c' is that it is not standard; for example, there is
no '-c' option in BSD variants (macOS). Using the 'stat' command is
a regression introduced in issue #616 to remove a dependency on the
format of the 'ls' output.