Changed to use predefined source
The implementation of egs_radionuclude_source is changed so
that it takes an input 'base source' instead of defining a
collimated or isotropic source.  This way you can use any
of the egs sources or a custom egs source to define the
location of the radionuclide in the simulation.  The change
in particular was made towards simulating targeted
radionuclide therapy.