Fix 3ddose scoring in egs_chamber application
Fix a bug where egs_chamber would crash when trying to generate a 3ddose
file from an XYZ geometry. This happened because the app pointer was not
defined for all of the calculation geometries, since they may skip the
geometry initialization that normally occurs. Now the application
pointer is set directly in the XYZ geometry.