Strip non-alphanumeric lead chars in PBS job name
Ensure that the PBS job name starts with an alphanumeric character
[0-9A-Za-z], following the PBS scheduler requirement. To avoid failed
jobs solely on the account of a bad job name, strip all leading
non-alphanumeric characters from the job name. Note that the egsinp
basename is not affected, this is strictly for the job name passed to
qsub via the -N option.