Fix egs-parallel leading character bug in job name
Fix a mistake in the previous fix for non-alpha numeric characters in
the job name in egs-parallel jobs resulting in a crash. The previous fix
removed non-alphanum characters from the start of the original job name,
but actually we needed to remove from the start of the trimmed 14
character job name.