EGS_Mesh: improve octree for overlapping elements
Rarely (~1.5% for a 600,000 element mesh), part of an element may be
contained within an octant but not counted as a bounded element because
none of the nodes or centroid are within the octant. The previous
implementation wrongly considered these points as outside (-1), while
the brute force method disagreed and found the correct element.

This issue was overcome by not immediately halting the search, but by
storing the size of the largest tetrahedral edge and comparing it to the
bounding box of the octant searched so far. If the search fails in the
most likely octant, its siblings are searched and then the parent's
siblings, etc, expanding the search area by moving up the octree. Only
when this "largest possible tetrahedron search" is completed, does the
octree conclude the point is indeed not inside the mesh.