EGS_Mesh: implement octree-accelerated hownear
Compared to both howfar and isWhere, hownear is much more expensive
even in the best case. We have to find the minimum distance to the mesh
in any direction, not just along one direction like howfar. So even
when a candidate element is found, we have to continue searching until
we conclude that no closer elements exist by testing against octree
bounding boxes.

If the minimum distance to an octant is larger than the current best
distance, we can skip searching that octant. Using this, we can avoid
checking much of the total surface mesh, especially if we find a good
candidate element early on.

Even so, many calls to hownear will harm the overall simulation
efficiency. This is currently the case for envelope simulations, since
hownear is called many times in the envelope before it reaches the
mesh. We're currently doing a full check for each hownear call, but
could potentially do a global bounding box check and return the
distance to the outer bounding box if the point is outside the mesh.
Hopefully this fudged distance won't impact correctness.