EGS_Mesh: return a lower bound in hownear_exterior
This change improves the overall simulation efficiency for envelope
electron simulations by avoiding an exhaustive check to find the real
minimum distance to the mesh. Instead, the distance to one of the
surface tree's bounding boxes may returned. For example, this caused the
efficiency of the MRCP_AM model in an air box with 1MeV photons, to
increase from ~400 to ~5000 histories per second.

This is allowed by the HOWNEAR spec (section 3.6 PIRS-701) which states:

"In complex geometries, the mathematics of HOWNEAR can become difficult
and sometimes almost impossible! If it is easier for the user to
compute some lower bound to the nearest distance, this could be used..."

Another alternative given in that document is setting the hownear result
to 0.0, forcing ELECTR to assume it is very close to a boundary at all
times, but that was found to be less efficient than this change.

So, given an input point to hownear, there are two possibilities:

1. If the point is outside the global bounding box, return the distance
to the closest point on the global bounding box and skip descending the
tree.

2. If the point is inside the global bounding box, descend the tree
until you reach a leaf node that contains the query point. A lower bound
can then be obtained by returning the minimum of the distance to the
bounding box sides and any boundary elements contained in the box.