EGS_Mesh: improve hownear_interior performance
Before this change, we were doing a full closest-triangle-point
calculation for query points inside the tetrahedron, which is overkill
because if a point is inside a tetrahedron, the shortest point will be
along a face normal. This translates to an approximate 10% performance
improvement for electron simulations.

This commit changes the minimum-interior calculation to a dot product
and 3D vector subtraction (i.e. distance to the triangle plane). We are
precalculating the face normals to further improve simulation
performance (5% faster compared to the baseline, calculating the normals
on the fly), but this comes with a fairly large increase in memory. Each
tetrahedron face normal has 3 double values, so the additional memory
per element is (4 x 3 x 8 =) 96 bytes. For the 8-million element MRCP
models, that is around 800 MB. We will offset this increase by no longer
storing the node coordinate data inline, but instead use offsets into a
node data vector.

This precalculation step will be somewhat of a waste for photon
simulations and if memory use becomes a problem it might be worthwhile
making this into a optional flag to be configured in the input file.