EGS_Mesh: pass to octree constructor, save memory
This change avoids copying node coordinate data over and over in the
octree constructor and instead looks it up using the EGS_Mesh. This
reduces the memory spike observed during initialization, which was
as high as 5GB but is now lowered to 4GB for the MRCP_AM model. The
steady state is around 2GB but should be lower after node coordinate
information is de-duplicated.