EGS_Mesh: fix howfar_interior for undershoots
Fix a major bug in howfar_interior. The previous implementation did not
ensure that the particle had to be "in front" of the tetrahedron face it
was intersecting with. So if the particle was slightly outside of the
tetrahedron as in the case of numerical undershoot, the wrong face would
be intersected. Here is a diagram of the situation:

      /\
* -> /  \
    /____\

The particle is numerically outside the tetrahedron, but EGS considers
it inside and we are required to calculate howfar as if it were inside
(see Bielajew's "HOWFAR and HOWNEAR: Geometry Modelling for Monte Carlo
Particle Transport"). So the near side must not be intersected. But the
previous implementation did not account for this and so sometimes a
particle travelling along one direction would be transported into a
neighbour element in the opposite direction. We now ensure the particle
must be on the "right" side of the tetrahedron face to intersect it.

In all there are three cases now handled in howfar_interior:

1. The position is inside the region: calculate the intersection without
   further complications.

2. The position is outside the region but will intersect one of the i
   interior faces: calculate the intersection, ignoring any backwards
   facing faces.

3. The position is outside the region and won't intersect any of the
   interior faces: return 0.0 as the howfar distance and set the new
   region number to match the position (even if it's not an immediate
   neighbour of the region in question).

    Case 1      |        Case 2       |        Case 3
                |                     |
      /\        |          /\         |          /\
     /  \       |         /  \        |         /  \
    /    \      |        /    \       |        /    \
   / * -> X     |  * -> /      X      |  <- * /      \
  /________\    |      /________\     |      /________\
                |                     |
 Intersection   |     Intersection    |    No intersection

Cases 1 and 2 are both handled by the same logic for now, but it
might be more efficient to use different equations for both (e.g.
ray-plane intersection algorithms for case 1).