Fix total dose in the getting started solutions
In the solution to one of the questions in the getting-started manual,
the total dose was calculated as the sum of dose to regions, instead of
appropriately weighted by the relative volume of each region. This was
corrected, as well as another typo in the instructions.

Reported by @mjukdisk in discussion #780.