Fix #803: window error from DYNJAWS in BEAMnrc GUI
Fix a GUI error that occurred in DYNJAWS jaw definitions when clicking
the button to calculate the jaw settings from the field size and SSD.
The button triggered a function that was duplicated with the same name
in SYNCJAWS, leading to the error. As the fix, the syncjaws version of
the function has been renamed. This bug has existed since SYNCJAWS was
added to the GUI.