EGS_Mesh: fix media offsets if there's other media
Use addMedium to get the correct offset for media initialization,
instead of starting from 0. If addMedium is called with a name already
in the media list, the corresponding offset is returned. Otherwise, the
media is added to the list and the new offset is returned.

Before this change, EGS_Mesh used to always define element media offsets
starting at 0. So if any media were defined before the mesh was created,
the mesh media offsets would be wrong.