Decrease minimum inter-CM gap from 0.01 to 1e-5 cm
Set the minimum distance between component modules in BEAMnrc to 1e-5 cm
(macro $MIN_GAP). Previously this minimum gap was set to 0.01 cm. Given
that the code is normally used in double-precision nowadays, there is no
point in leaving such a large value as the default.

This change follows for looking into issue #773.