Add new leaf orientation option for synchdmlc
Add $SYNCHDMLC_ORIENT options to flip leaf numbering and cross-sections
with respect to the axis perpendicular to the opening axis, so that the
MLC coordinate system matches that in the manufacturer (Varian) specs
for 120MLC and HD120.

This reflection reverses the order of leaf numbering, hence leaf opening
sequences are automatically "flipped" so that the opening coordinate
applied to leaf i is instead applied to leaf N-i+1, where N is the total
number of leaves. See the updated BEAMnrc manual for more details.

Also, add new ORIENT=2,3 options to SYNCHDMLC gui