Fix buffer overflow from scoring vacuum media dose
Vacuum has a special medium number -1, while all other media get offsets
from 0 to nmedia. Before this change, there was a buffer overflow if
media dose scoring was used with a geometry that had vacuum as one of
the media. The vector massM is indexed by medium number, but for any
vacuum regions, it would read and write memory at the memory location
one before the buffer bounds: massM[-1].

Skipping the vacuum region masses shouldn't matter, as vacuum dose is
always zero. massM is only used later on with non-vacuum media, so
skipping vacuum regions is the right choice instead of an alternative
like appending the vacuum mass to massM.