Fix branch logging and code comparisons
Remove the git branch information and provide the commit hash only. The
branch name is just a label that can point to any commit, so it remains
equivocal. This also solves the issue with long branch names where the
fortran code would fail to compile because of exceeding the line length.

Also, fix the comparisons between egs_home and HEN_HOUSE/user_code apps,
so that it doesn't report a difference when the HEN_HOUSE version
doesn't exist.