Add gamma spectrometry application egs_gammasec
Add an application for calculating gamma spectrometry related
quantities, like detector efficiency and coincidence summing
corrections. The egs_radionuclide_source was changed slightly to allow
for the calculations - it now returns the number of decays from
getNextParticle, instead of the number of source particles. This does
not change the normalization when using the source, but apps specially
designed to use current_case and last_case for the radionuclide source
may be affected.