Fix naive EGS_Triangle_Mesh isWhere implementation
Instead of a less-than comparison of the minimum interior and exterior
distances, less-or-equal is required to correctly implement isWhere.

Rays intersecting the mesh exactly on the corner can lead to a case
where the minimum interior distance and minimum exterior distance are
equal. If this occurs, it should be counted as outside the mesh,
otherwise, extra dose is wrongly attributed to the mesh region.