Fix egs_ensdf bug for ensdf comments in nndc files
Fix a bug where some ensdf files from the nndc caused a segmentation
fault for egs_radionuclide source. This was due to comment lines
following gamma records in the file. The code did not check to see if a
normalization record existed before trying to use it, and this has been
fixed.