Add dynamic shape that moves during a simulation
Add a shape that handles dynamic motion, synchronized with the motion
of geometries and sources.

Add dynamic shape doxygen documentation, astyle

Fix flickering of dynamic particle tracks

Fix flickering that occurred when dragging the time slider with the
mouse. Particle tracks would flash on the screen from a longer time
range than the selected window.

Fix memory leak in dynamic geometry and shape

Fix motion for shapes lacking source direction

Adjust the default viewer time window

Change the time window default to 1% of the simulation duration for
tracks visualization. Also add a new input in egs_view for the number of
time steps to take in the animation. This allows for better control of
the visual speed of the motion.

Remove shape motion control point numbering syntax

Adjust coding style to EGSnrc standard with astyle

Fix missing updateNumTimeSteps function

Somehow a function got deleted during some git manipulation, so this is
just adding it back it.

Set the slider based on the number of time steps