Add time into the standard ptracks format
The ptracks file will now start with "include time index=" and then a
boolean value. egs_view has been updated to support this. This avoids
creating a new .syncptracks file extension, that was originally done in
this implementation. Also changed the default behaviour for
egs_track_scoring so that time index scoring is automatically turned on
when dynamic geometries or sources are used, if "include time index" is
excluded from the input file.