Test IRL values with BEAMnrc simulations
Follow-up on previous commit: Why did the bug not appear during
standalone BEAMnrc runs?

During direct BEAMnrc simulations, when IRL isn't set in the block to
split EII-generated secondary photons, it takes value 2, the first
region of the geometry, set during previous calls to AUSGAB.

When using BEAMnrc as a library with other applications, the memory
location IRL is pointing to can change, taking an arbitrary number. In
my case it was a very large integer.