Harmonize calc types in g
First change:
------------

Calculation type 1 did not incorporate the changes implemented in 2017
to remove kerma calculation dependency to transport and production
thresholds ECUT and AE, as well as the inclusion of electron impact
ionization in the mutr calculation step.

As a consequence, muen and mutr values calculated with either
calculation type do not match. The reason for the mismatch being that
there is no distinction of the energy deposited by sub-threshold events
between Auger electrons and fluorescent photons, with the first being
part of kerma while the second not.

This fix corrects this by mainly adding AUSGAB calls before and after
the routines for photo-effect, Compton scattering and pair production.

Notably, current NIST muen values are only matched without this fix,
suggesting that there might be a need to update those values.

Second change:
-------------

Add option to reproduce NIST values by setting input key below to 'no'

CLASSIFY SUB-THRESHOLD RELAXATIONS = NO

in which case no calls to AUSGAB are made during photon interactions in
the mutr-calculating step of a type 1 calculation. MOreover, for type 0
calculations a check is performed in ausgab to apply or not the
classification schema or not, based on the value of the logical variable
classify_sub_threshold_relaxations. Default set to `yes`.