Fix th-234 lnhb ensdf support
Fix the parsing of Th-234 in the LNHB ensdf file. This change checks for
repeated levels with the same energy and different spin-parity that are
included in the ensdf file to represent an isomeric transition. The low
probability level is neglected now, because no probability is provided
for the transition. Previously, both levels contained high probability
beta decays towards them and were treated equivalently, leading to
an incorrect decay scheme.