Fix phsp scoring outside the geometry on entry
Fix a bug where particles were scored in the phase-space even though
they had left the overall geometry. This was occurring even when the
particles were only meant to be scored on entry, if the region of the
phase-space scoring geometry happened to be defined outside the actual
geometry used in the simulation (like an infinite plane).