Add mesh support to egs_dose_scoring
Automatically determine the mass for calculating dose for both mesh and
XYZ geometries, when the egs_dose_scoring ausgab object is set to output
a dose file. This allows for dose to each medium to be reported
correctly for XYZ and mesh geometries.

Also fix a bug where dose to 3ddose files would be wrong if the XYZ
geometry was transformed.