Fix #1300: egs_view crash due to bug in dyngeom
Fix a segmentation fault that was occurring in egs_view due to accessing
memory that was not initialized in the geometry. This occurred during a
call to containsDynamic(), which in the composite geometries was missing
a check on whether a geometry existed before propogating the
containsDynamic() call. The result was a segmentation fault in egs_view
for specific use cases of a CD geometry.