Enable region ranges and default-all for labels
Add the ability to apply a label to all regions in the local geometry by
default when no region number is specified. Add parsing of region ranges
specified with a hyphen, as in "5-12", to label all regions in the
range. Both multiple region numbers and ranges can be specified in the
same "set label" input. This is handy for geometries containing a large
number of regions. Previously, all region numbers had to be specified
individually. Also, improve tokenization and parsing robustness, as well
as error messages.